+++ packages/fcl/src/current-user/exec-service/strategies/utils/pop.js
@@ -1,9 +1,11 @@
 import {renderPop} from "./render-pop"
 import {serviceEndpoint} from "./service-endpoint"
 
-const CLOSE_EVENT = "FCL:FRAME:CLOSE"
-const READY_EVENT = "FCL:FRAME:READY"
-const RESPONSE_EVENT = "FCL:FRAME:RESPONSE"
+const CLOSE_EVENT = "FCL:VIEW:CLOSE"
+const READY_EVENT = "FCL:VIEW:READY"
+const RESPONSE_EVENT = "FCL:VIEW:RESPONSE"
+
+export const _ = e => typeof e === "string" && e.toLowerCase()
 
 const noop = () => {}
 
@@ -32,18 +34,23 @@ export function pop(service, opts = {}) {
     try {
       if (typeof e.data !== "object") return
       if (IGNORE.has(e.data.type)) return
-      if (e.data.type === CLOSE_EVENT) close()
-      if (e.data.type === READY_EVENT) onReady(e, {send, close})
-      if (e.data.type === RESPONSE_EVENT) onResponse(e, {send, close})
+      if (_(e.data.type) === _(CLOSE_EVENT)) close()
+      if (_(e.data.type) === _(READY_EVENT)) onReady(e, {send, close})
+      if (_(e.data.type) === _(RESPONSE_EVENT)) onResponse(e, {send, close})
       onMessage(e, {send, close})
 
       // Backwards Compatible
-      if (e.data.type === "FCL::CHALLENGE::RESPONSE") {
+      if (_(e.data.type) === _("FCL:FRAME:READY")) onReady(e, {send, close})
+      if (_(e.data.type) === _("FCL:FRAME:RESPONSE"))
+        onResponse(e, {send, close})
+      if (_(e.data.type) === _("FCL:FRAME:CLOSE")) close()
+      //
+      if (_(e.data.type) === _("FCL::CHALLENGE::RESPONSE")) {
         onResponse(e, {send, close})
       }
-      if (e.data.type === "FCL::AUTHZ_READY") onReady(e, {send, close})
-      if (e.data.type === "FCL::CHALLENGE::CANCEL") close()
-      if (e.data.type === "FCL::CANCEL") close()
+      if (_(e.data.type) === _("FCL::AUTHZ_READY")) onReady(e, {send, close})
+      if (_(e.data.type) === _("FCL::CHALLENGE::CANCEL")) close()
+      if (_(e.data.type) === _("FCL::CANCEL")) close()
     } catch (error) {
       console.error("Popup Callback Error", error)
       close()